requirements.txt
playwright==1.43.0
pandas==2.2.2
python-dotenv
supabase==2.0.3
pytz
python-dateutil
Flask==3.0.0
gunicorn==21.2.0